change on declaration

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -8,7 +8,7 @@
  */
 size_t print_listint(const listint_t *h)
 {
-int the_counter = 1;
+size_t the_counter = 1;
 
 printf("%d\n", h->n);
 if (h->next != NULL)
